// src/app/models/forum.model.ts
export interface Notification {
id: string;
message: string;
topicId: string;
topicTitle: string;
isRead: boolean;
createdAt: Date | string;
author?: {
name: string;
avatar?: string;
};
}